+++ src/idunno.Bluesky/Graph/ListWithMembership.cs
@@ -8,7 +8,7 @@ namespace idunno.Bluesky.Graph;
 /// <summary>
 /// A list and an optional list item indicating membership of a target user to that list.
 /// </summary>
-public record ListWithMembership: View
+public record ListWithMembership : View
 {
     /// <summary>
     /// Creates a new instance of <see cref="ListWithMembership"/>.
@@ -30,7 +30,7 @@ public record ListWithMembership: View
     public ListView List { get; init; }
 
     /// <summary>
-    /// Gests the list item indicating membership of a target user to the list.
+    /// Gets the list item indicating membership of a target user to the list.
     /// </summary>
     [JsonInclude]
     public ListItemView? ListItem { get; init; }